Adding plotter

+++ single_pulse_ml/moose-script.py
@@ -85,8 +85,8 @@ if plot:
     elif train_set is False and run_predict is True:
 
         figname = './single_pulse_ml/plots/%s_test.png' % array_name 
-        pred_name = plot_tools.get_title(y_pred, target_names)
-        true_name = plot_tools.get_title(y_pred, target_names)
+        pred_name = plot_tools.get_title(y_pred.astype(int), target_names)
+        true_name = plot_tools.get_title(y_pred.astype(int), target_names)
         print pred_name
         prediction_titles = ['predicted: %s\ntrue:      %s' % (pred_name[ii], true_name[ii])     
                         for ii in range(len(pred_name))]
